@@ -1,18 +1,19 @@
 from typing import cast
+import uuid
 
 from langgraph.graph import StateGraph, END
 
 from madd.core.state import DebateState
-from madd.agents import (
-    generate_profile,
-    generate_turn,
-    evaluate_round,
-    verify_claims,
+from madd.core.schemas import (
+    Clause, ClauseStatus, TreatyDraft, DebateMessage, ProposedClause
 )
+from madd.stores.profile_store import ensure_profile
+from madd.agents.country import generate_turn
+from madd.agents.judge import evaluate_round
+from madd.agents.verifier import verify_claims
 
 
 def build_graph() -> StateGraph:
-
     graph = StateGraph(DebateState)
     
     graph.add_node("ensure_profiles", _ensure_profiles)
@@ -23,15 +24,12 @@ def build_graph() -> StateGraph:
     graph.add_node("judge", _judge)
     graph.add_node("finalize_report", _finalize_report)
     
-    # Define edges
     graph.set_entry_point("ensure_profiles")
     graph.add_edge("ensure_profiles", "opening_statements")
-    graph.add_edge("opening_statements", "negotiate_round")
-    graph.add_edge("negotiate_round", "compile_treaty")
+    graph.add_edge("opening_statements", "compile_treaty")
     graph.add_edge("compile_treaty", "verify")
     graph.add_edge("verify", "judge")
     
-    # Conditional edge: continue or finalize
     graph.add_conditional_edges(
         "judge",
         _should_continue,
@@ -40,129 +38,154 @@ def build_graph() -> StateGraph:
             "finalize": "finalize_report",
         }
     )
+    graph.add_edge("negotiate_round", "compile_treaty")
     graph.add_edge("finalize_report", END)
     
     return graph.compile()
 
 
-# =============================================================================
-# Node Implementations
-# =============================================================================
-
 def _ensure_profiles(state: DebateState) -> dict:
-
     scenario = state["scenario"]
-    profiles = dict(state["profiles"]) # Copy
+    profiles = {}
     
-    print(f"--- Node: Ensure Profiles ---")
+    print(f"[ensure_profiles] Loading/generating profiles for {len(scenario.countries)} countries")
     
     for country in scenario.countries:
-        if country not in profiles:
-            print(f"Generating profile for: {country}")
-            profiles[country] = generate_profile(country, scenario.description)
-            
+        print(f"  - {country}...")
+        profiles[country] = ensure_profile(country, scenario.description)
+    
     return {"profiles": profiles}
 
 
 def _opening_statements(state: DebateState) -> dict:
-    """Collect opening statements from all countries."""
     scenario = state["scenario"]
-    messages = []
-    
-    print(f"--- Node: Opening Statements (Round 1) ---")
-    
-    for country in scenario.countries:
-        # For opening statements, we can set round=1 context
-        # Ideally, we update the state round before this loop, 
-        # but since 'opening_statements' implies start, we force it.
-        # However, the node return value updates the state.
-        
-        # We need to pass a temp state with round=1 to the agent?
-        # Or just let the agent see round=0 and know it's opening?
-        # Let's verify: State has 'round'. We should probably increment round to 1 here?
-        pass 
-
-    # Better approach: Iterate countries, generate turns.
-    # Since this is "opening", it's effectively Round 1.
+    print(f"[opening_statements] Round 1")
     
     new_messages = []
+    temp_state = dict(state)
+    temp_state["round"] = 1
+    
     for country in scenario.countries:
-        print(f"{country} is speaking...")
-        # We need to temporarily mock round=1 for the agent if it's currently 0
-        context_state = state.copy()
-        context_state["round"] = 1 
-        
-        msg = generate_turn(context_state, country)
+        print(f"  - {country} speaking...")
+        msg = generate_turn(temp_state, country)
+        msg.round_number = 1
         new_messages.append(msg)
-        
+        temp_state["messages"] = list(state.get("messages", [])) + new_messages
+    
     return {"messages": new_messages, "round": 1}
 
 
 def _negotiate_round(state: DebateState) -> dict:
-    """Run a negotiation round with all countries."""
     scenario = state["scenario"]
-    current_round = state["round"] + 1 # Increment for this new round
-    
-    print(f"--- Node: Negotiate Round {current_round} ---")
+    current_round = state["round"] + 1
+    print(f"[negotiate_round] Round {current_round}")
     
     new_messages = []
-    # In a real debate, order might shuffle or be sequential.
-    # Currently simple iteration.
+    temp_state = dict(state)
+    temp_state["round"] = current_round
+    
     for country in scenario.countries:
-        print(f"{country} is speaking...")
-        
-        # Update context to reflect this new round
-        context_state = state.copy()
-        context_state["round"] = current_round
-        
-        # Also need to append previous messages from THIS round to context?
-        # Our state["messages"] only gets updated at the END of the node execution in LangGraph (usually).
-        # To allow agents to react to each other WITHIN the round, we might need a sub-loop 
-        # or append to context_state["messages"].
-        
-        # For MVP: Simultaneous turns (agents don't see each other's messages in the SAME round until next node).
-        # Or we append locally:
-        context_state["messages"] = list(state["messages"]) + new_messages
-        
-        msg = generate_turn(context_state, country)
+        print(f"  - {country} speaking...")
+        temp_state["messages"] = list(state.get("messages", [])) + new_messages
+        msg = generate_turn(temp_state, country)
+        msg.round_number = current_round
         new_messages.append(msg)
-        
+    
     return {"messages": new_messages, "round": current_round}
 
 
 def _compile_treaty(state: DebateState) -> dict:
-
-    # MVP: Just copy existing treaty for now. 
-    # Real logic: Parse 'clause_votes' from messages and update clause statuses.
-    print(f"--- Node: Compile Treaty ---")
+    print(f"[compile_treaty] Processing clauses")
     
-    # TODO: Implement complex clause voting logic.
-    # For now, pass through.
-    return {}
+    current_round = state["round"]
+    messages = state.get("messages", [])
+    treaty = state.get("treaty") or TreatyDraft()
+    countries = state["scenario"].countries
+    existing_clause_ids = {c.id for c in treaty.clauses}
+    
+    round_messages = [m for m in messages if m.round_number == current_round]
+    
+    new_clauses = []
+    for msg in round_messages:
+        for proposed in msg.proposed_clauses:
+            clause_id = f"C{len(treaty.clauses) + len(new_clauses) + 1}"
+            new_clauses.append(Clause(
+                id=clause_id,
+                text=proposed.text,
+                proposed_by=msg.country,
+                status=ClauseStatus.PROPOSED,
+                proposed_round=current_round,
+                supporters=[msg.country],
+                objectors=[],
+            ))
+    
+    all_clauses = list(treaty.clauses) + new_clauses
+    
+    for clause in all_clauses:
+        if clause.status != ClauseStatus.PROPOSED:
+            continue
+        
+        for msg in round_messages:
+            if msg.country == clause.proposed_by:
+                continue
+            
+            vote = msg.clause_votes.get(clause.id, "")
+            if vote == "support" and msg.country not in clause.supporters:
+                clause.supporters.append(msg.country)
+            elif vote == "oppose" and msg.country not in clause.objectors:
+                clause.objectors.append(msg.country)
+    
+    for clause in all_clauses:
+        if clause.status != ClauseStatus.PROPOSED:
+            continue
+        
+        support_count = len(clause.supporters)
+        oppose_count = len(clause.objectors)
+        total = len(countries)
+        
+        if support_count > total / 2:
+            clause.status = ClauseStatus.ACCEPTED
+            clause.resolved_round = current_round
+        elif oppose_count > total / 2:
+            clause.status = ClauseStatus.REJECTED
+            clause.resolved_round = current_round
+    
+    updated_treaty = TreatyDraft(
+        title=treaty.title or f"Treaty on {state['scenario'].name}",
+        preamble=treaty.preamble,
+        clauses=all_clauses,
+    )
+    
+    return {"treaty": updated_treaty}
 
 
 def _verify(state: DebateState) -> dict:
-
-    print(f"--- Node: Verify ---")
-    audit_findings = verify_claims(state)
-    return {"audit": audit_findings}
+    print(f"[verify] Checking claims")
+    try:
+        findings = verify_claims(state)
+        return {"audit": findings}
+    except Exception as e:
+        print(f"  Verification error: {e}")
+        return {"audit": []}
 
 
 def _judge(state: DebateState) -> dict:
-    print(f"--- Node: Judge ---")
-    scorecard = evaluate_round(state)
-    return {"scorecards": [scorecard]}
+    print(f"[judge] Evaluating round {state['round']}")
+    try:
+        scorecard = evaluate_round(state)
+        return {"scorecards": [scorecard]}
+    except Exception as e:
+        print(f"  Judge error: {e}")
+        from madd.core.schemas import RoundScorecard
+        return {"scorecards": [RoundScorecard(round_number=state["round"])]}
 
 
 def _finalize_report(state: DebateState) -> dict:
-    print(f"--- Node: Finalize Report ---")
-    # MVP: Just print done
-    print("Debate concluded.")
+    print(f"[finalize_report] Debate complete after {state['round']} rounds")
     return {}
 
 
 def _should_continue(state: DebateState) -> str:
-    """Determine if debate should continue or finalize."""
     current_round = state.get("round", 1)
     max_rounds = state.get("max_rounds", 3)
     
