@@ -1,5 +1,3 @@
-"""Core module exports."""
-
 from madd.core.schemas import (
     Citation,
     EconomicData,
@@ -7,34 +5,42 @@ from madd.core.schemas import (
     CountryStrategy,
     CountryProfile,
     Clause,
+    ClauseStatus,
     TreatyDraft,
+    CountryScore,
     RoundScorecard,
+    AuditSeverity,
     AuditFinding,
+    ProposedClause,
     DebateMessage,
 )
 from madd.core.scenario import Scenario, AgendaItem, load_scenario
-from madd.core.state import DebateState
-from madd.core.config import Settings, get_settings
+from madd.core.state import DebateState, create_initial_state
+from madd.core.config import Settings, get_settings, current_year
+from madd.core.graph import build_graph
 
 __all__ = [
-    # Schemas
     "Citation",
     "EconomicData",
     "CountryFacts",
     "CountryStrategy",
     "CountryProfile",
     "Clause",
+    "ClauseStatus",
     "TreatyDraft",
+    "CountryScore",
     "RoundScorecard",
+    "AuditSeverity",
     "AuditFinding",
+    "ProposedClause",
     "DebateMessage",
-    # Scenario
     "Scenario",
     "AgendaItem",
     "load_scenario",
-    # State
     "DebateState",
-    # Config
+    "create_initial_state",
     "Settings",
     "get_settings",
+    "current_year",
+    "build_graph",
 ]
